pyproject.toml
[project]
name = "vladpy-telegram-ro-bot"
description = "Telegram Romanian translator Bot, by vladpy"
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Vladislav Pyslaru", email = "vlpyslaru+tgrobot@gmail.com"},
]
requires-python = ">=3.11"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Development Status :: 1 - Planning",
]
dependencies = [
    "click~=8.1",
    "python-telegram-bot~=21.1",
    "aiohttp~=3.9",
]
dynamic = [
    "version",
]

[project.urls]
homepage = "https://github.com/vladpy8/telegram-ro-bot"
repository = "https://github.com/vladpy8/telegram-ro-bot"


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pylint]

[tool.pylint.format]
indent-string="\t"
max-line-length=120

[tool.pylint.design]
min-public-methods=1
max-args=10

[tool.pylint.string]
check-quote-consistency=true

[tool.pylint.message-control]
disable=[
    "broad-exception-caught",
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring",
]


[tool.pyright]
typeCheckingMode="strict"


[tool.mypy]
warn_redundant_casts=true
warn_unused_ignores=true
warn_unreachable=true

allow_redefinition=true
strict_equality=true
strict=true


[tool.pdm]
version = { source = "file", path = "src/vladpy_telegram_ro_bot/__version__.py" }

[tool.pdm.dev-dependencies]
dev = [
    "pylint~=3.1",
    "mypy~=1.10",
    "yapf~=0.40",
    "pyright~=1.1",
]
